[Refactor]: CMCException 경로 변경 적용
