+++ src/main/java/com/sctk/cmc/domain/Designer.java
@@ -1,7 +1,6 @@
 package com.sctk.cmc.domain;
 
-import com.sctk.cmc.exception.CMCException;
-import com.sctk.cmc.exception.ResponseStatus;
+import com.sctk.cmc.common.exception.CMCException;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -10,7 +9,7 @@ import javax.persistence.*;
 import java.util.ArrayList;
 import java.util.List;
 
-import static com.sctk.cmc.exception.ResponseStatus.*;
+import static com.sctk.cmc.common.exception.ResponseStatus.*;
 
 @Getter
 @NoArgsConstructor
