+++ src/main/java/com/sctk/cmc/service/AuthServiceImpl.java
@@ -2,7 +2,7 @@ package com.sctk.cmc.service;
 
 import com.sctk.cmc.domain.Designer;
 import com.sctk.cmc.domain.Member;
-import com.sctk.cmc.exception.CMCException;
+import com.sctk.cmc.common.exception.CMCException;
 import com.sctk.cmc.service.abstractions.AuthService;
 import com.sctk.cmc.service.abstractions.DesignerService;
 import com.sctk.cmc.service.abstractions.MemberService;
@@ -10,7 +10,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
-import static com.sctk.cmc.exception.ResponseStatus.*;
+import static com.sctk.cmc.common.exception.ResponseStatus.*;
 
 @RequiredArgsConstructor
 @Service
